Renamed "SamSharpTest" to "SamSharpCli", so that we can set up a new project that actually defines automated tests.
